refactor: component -> util 패키지 이름 변경
